git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/anatomist/anatomist-free/trunk@70147 613412ce-1361-4b7a-93bf-a3c75af1bae4

+++ src/lib/reference/Transformation.cc
@@ -282,7 +282,7 @@ void Transformation::addMotionToHistory(const Motion & motion)
 		_motionHistory.clear();
 	}
 	else if (_motionHistoryIndex >= 0 &&
-		_motionHistoryIndex < (_motionHistory.size() - 1))
+                 static_cast<size_t>(_motionHistoryIndex) < (_motionHistory.size() - 1))
 	{
 		_motionHistory.erase(_motionHistory.begin() + _motionHistoryIndex + 1,
 							 _motionHistory.end());
@@ -308,7 +308,7 @@ void Transformation::undo()
 
 void Transformation::redo()
 {
-	if (_motionHistoryIndex == (_motionHistory.size() - 1) ||
+  if (static_cast<size_t>(_motionHistoryIndex) == (_motionHistory.size() - 1) ||
 		_motionHistory.empty())
 	{
 		return;
